Seasons añadidas faltan episodes

+++ laravel/routes/web.php
@@ -16,7 +16,6 @@ Route::middleware([
     'auth:sanctum',
     config('jetstream.auth_session'),
     'verified',
-    'role:admin|pay'
 ])->group(function () {
     $controller_path = 'App\Http\Controllers';
 
@@ -46,13 +45,11 @@ Route::resource('series.seasons', SeasonController::class)->middleware(['auth'])
  * 
  */
 
-Route::middleware(['auth:sanctum', config('jetstream.auth_session'), 'verified', 'role:admin|pay'])->group(function () {
     Route::get('/movies', [MovieController::class, 'index'])->name('movies.index');
     Route::get('/movies/{movie}', [MovieController::class, 'show'])->name('movies.show');
 
     Route::get('/series', [SerieController::class, 'index'])->name('series.index');
     Route::get('/series/{serie}', [SerieController::class, 'show'])->name('series.show');
-});
 
 /**
  * 
@@ -67,11 +64,11 @@ Route::middleware(['auth:sanctum', config('jetstream.auth_session'), 'verified',
      * 
      */
 
-    Route::post('/movies', [MovieController::class, 'store'])->name('movies.store');
-    Route::get('/movies/create', [MovieController::class, 'create'])->name('movies.create');
-    Route::put('/movies/{movie}', [MovieController::class, 'update'])->name('movies.update');
-    Route::delete('/movies/{movie}', [MovieController::class, 'destroy'])->name('movies.destroy');
-    Route::get('/movies/{movie}/edit', [MovieController::class, 'edit'])->name('movies.edit');
+    Route::post('/admin/movies', [MovieController::class, 'store'])->name('movies.store');
+    Route::get('/admin/movies/create', [MovieController::class, 'create'])->name('movies.create');
+    Route::put('/admin/movies/{movie}', [MovieController::class, 'update'])->name('movies.update');
+    Route::delete('/admin/movies/{movie}', [MovieController::class, 'destroy'])->name('movies.destroy');
+    Route::get('/admin/movies/{movie}/edit', [MovieController::class, 'edit'])->name('movies.edit');
     /**
      * 
      * SERIES
@@ -79,9 +76,9 @@ Route::middleware(['auth:sanctum', config('jetstream.auth_session'), 'verified',
      */
 
     Route::post('/series', [SerieController::class, 'store'])->name('series.store');
-    Route::get('/series/create', [SerieController::class, 'create'])->name('series.create');
-    Route::put('/series/{serie}', [SerieController::class, 'update'])->name('series.update');
-    Route::delete('/series/{serie}', [SerieController::class, 'destroy'])->name('series.destroy');
+    Route::get('/admin/series/create', [SerieController::class, 'create'])->name('series.create');
+    Route::put('/admin/series/{serie}', [SerieController::class, 'update'])->name('series.update');
+    Route::delete('/admin/series/{serie}', [SerieController::class, 'destroy'])->name('series.destroy');
     Route::get('/series/{serie}/edit', [SerieController::class, 'edit'])->name('series.edit');
     /**
      * 
@@ -91,12 +88,12 @@ Route::middleware(['auth:sanctum', config('jetstream.auth_session'), 'verified',
 });
 
 Route::get('/series/{serie}/seasons', [SeasonController::class, 'index'])->name('seasons.index');
-Route::get('/series/{serie}/seasons/create', [SeasonController::class, 'create'])->name('seasons.create');
-Route::post('/series/{series}/seasons', [SeasonController::class, 'store'])->name('series.seasons.store');
+Route::get('/admin/series/{serie}/seasons/create', [SeasonController::class, 'create'])->name('seasons.create');
+Route::post('/admin/series/{serie}/seasons', [SeasonController::class, 'store'])->name('series.seasons.store');
 Route::get('/seasons/{season}', [SeasonController::class, 'show'])->name('seasons.show');
-Route::get('/seasons/{season}/edit', [SeasonController::class, 'edit'])->name('seasons.edit');
-Route::put('/seasons/{season}/update', [SeasonController::class, 'update'])->name('seasons.update');
-Route::delete('/seasons/{season}/destroy', [SeasonController::class, 'destroy'])->name('seasons.destroy');
+Route::get('/admin/seasons/{season}/edit', [SeasonController::class, 'edit'])->name('seasons.edit');
+Route::put('/admin/seasons/{season}/update', [SeasonController::class, 'update'])->name('seasons.update');
+Route::delete('/admin/seasons/{season}/destroy', [SeasonController::class, 'destroy'])->name('seasons.destroy');
 
 /**
  *
